Java_PlaneManagementProgram Project
In this project, we used Java to create a plane ticket process. This is where the user can book their seat, cancel their seat, see which seat or row is available and they can add their details which are stored. The user also has the option to print out the seating plan when they have booked their seat or before booking.