Python and Graphics.py
In this project, we used Python and Graphics.py to display a histogram of the student's exam results. We had 4 categories for the student's results: Progress, Progress (module trailer), Do not progress - module retriever and Exculde. We also created a txt called DataList which stores all the data the user enters.