Java Plane Management Program
In this project, we used Java to create a plane ticket process. Where a new company is managing a private plane and asked to create a Java program to manage and track seat reservations. This project involved designing and implementing this program. There are four rows of seats: A, B, C, and D and these rows have varying seat capacities (either 12 or 14). The seats are coded according to their row and seat number. For example, seat B2 is marked with an 'X' and costs £200.