Health Center
This was a Lab practical exam in which we were provided a template code alongside a UML design. We had to add a Nurse class, abstract method called deleteStaffMember which we had to further override the method by asking the user to insert the staff ID of the member they want to delete. It should print the name, surname, and whether they are a nurse, doctor or receptionist. We had to modify the run menu and sort the arrayList by using collections and implementing a Comparable interface. Additionally, we had to add a column, add an event handler and create a unit test using the JUnit Framework.