Web Design and Development
This was a group project where we selected the topic "Responsible Consumption & Production" from the UN website and created a website based on it. My role as Student 2 involved ensuring cohesive styling by providing the CSS stylesheet for the entire team, as each member was responsible for creating different pages of the website. I also developed a feedback form that allowed customers to enter their details and rate their satisfaction with the website. Upon submission, a summary of the form would appear at the bottom of the page, enabling users to make any edits before final submission. Once finalized, the form displayed a confirmation email box with the completed summary. Additionally, I was responsible for designing the homepage, which included interactive buttons and a navigation bar to facilitate seamless navigation throughout the website.